src/fabrik-service/requirements.txt
Flask==2.3.3
redis==5.0.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-exporter-otlp-proto-http==1.20.0
